trivial: Split cli.py into smaller files
