+++ tests/cli/expected/help_backlinks.txt
@@ -5,7 +5,7 @@ List pages that link to a Wikipedia page.
 TITLE is the Wikipedia page title.
 
 \b Examples:     wikipedia-api backlinks "Python (programming language)"
-wikipedia-api backlinks "Python (programming language)" --json r
+wikipedia-api backlinks "Python (programming language)" --json
 
 Options:
 -l, --language TEXT             Language edition of Wikipedia (e.g. en, cs,
