+++ tests/cli/expected/help_categories.txt
@@ -5,7 +5,7 @@ List categories for a Wikipedia page.
 TITLE is the Wikipedia page title.
 
 \b Examples:     wikipedia-api categories "Python (programming language)"
-wikipedia-api categories "Python (programming language)" --json r
+wikipedia-api categories "Python (programming language)" --json
 
 Options:
 -l, --language TEXT             Language edition of Wikipedia (e.g. en, cs,
