+++ tests/cli/expected/help_categorymembers.txt
@@ -8,7 +8,7 @@ Use --max-level to recursively list subcategory members.
 
 \b Examples:     wikipedia-api categorymembers "Category:Physics"
 wikipedia-api categorymembers "Category:Physics" --max-level 1
-wikipedia-api categorymembers "Category:Physics" --json r
+wikipedia-api categorymembers "Category:Physics" --json
 
 Options:
 --max-level INTEGER             Maximum depth for recursive category member
