+++ tests/cli/expected/help_links.txt
@@ -5,7 +5,7 @@ List pages linked from a Wikipedia page.
 TITLE is the Wikipedia page title.
 
 \b Examples:     wikipedia-api links "Python (programming language)"
-wikipedia-api links "Python (programming language)" --json r
+wikipedia-api links "Python (programming language)" --json
 
 Options:
 -l, --language TEXT             Language edition of Wikipedia (e.g. en, cs,
