+++ tests/cli/expected/help_main.txt
@@ -12,7 +12,7 @@ Examples: \b     wikipedia-api summary "Python (programming language)"
 wikipedia-api links "Python (programming language)" --language cs
 wikipedia-api categories "Python (programming language)" --json
 wikipedia-api coordinates "Mount Everest"     wikipedia-api geosearch
---coord "51.5074|-0.1278"     wikipedia-api search "Python programming" r
+--coord "51.5074|-0.1278"     wikipedia-api search "Python programming"
 
 Options:
 --version   Show the version and exit.
