+++ tests/cli/expected/help_random.txt
@@ -4,7 +4,7 @@ Get random Wikipedia pages.
 
 \b Examples:     wikipedia-api random     wikipedia-api random --limit 5
 wikipedia-api random --filter-redirect all --json     wikipedia-api random
---min-size 1000 --max-size 10000     wikipedia-api random --language de r
+--min-size 1000 --max-size 10000     wikipedia-api random --language de
 
 Options:
 --limit INTEGER                 Number of random pages to return.  [default:
