+++ tests/cli/expected/langlinks_fr.txt
@@ -1,3 +1,4 @@
+abstract: Q90 (https://abstract.wikipedia.org/wiki/Q90)
 ace: Paris (https://ace.wikipedia.org/wiki/Paris)
 ady: Париж (https://ady.wikipedia.org/wiki/%D0%9F%D0%B0%D1%80%D0%B8%D0%B6)
 af: Parys (https://af.wikipedia.org/wiki/Parys)
