+++ tests/cli/expected/random_combined.txt
@@ -1,10 +1,10 @@
 [
 {
-"title": "Let Go (Avril Lavigne album)",
-"pageid": 276783
+"title": "Rafał Śliż",
+"pageid": 21153094
 },
 {
-"title": "WTA Atlanta",
-"pageid": 42302530
+"title": "List of Billboard China V Chart number-one videos of 2017",
+"pageid": 55565370
 }
 ]
