+++ tests/cli/expected/random_default.txt
@@ -1,1 +1,1 @@
-S.C. Dacy
+The Voice UK series 9
