+++ tests/cli/expected/random_filter_redirects.txt
@@ -1,3 +1,3 @@
-Melville Land
-The Tolkien Society
-Nashe Radio
+Glory Kickboxing
+1898 Penn Quakers football team
+Global Telecom & Technology, Inc.
