+++ tests/cli/expected/random_filter_redirects_json.txt
@@ -1,6 +1,6 @@
 [
 {
-"title": "Digambara-Svetambara split",
-"pageid": 79737265
+"title": "Omohyoid",
+"pageid": 894703
 }
 ]
