+++ tests/cli/expected/random_json.txt
@@ -1,14 +1,14 @@
 [
 {
-"title": "Rosina (ship)",
-"pageid": 62813897
+"title": "Gedney Hill",
+"pageid": 1573312
 },
 {
-"title": "Jeffrey Blackburn",
-"pageid": 74674689
+"title": "Arthur B. Metzner",
+"pageid": 18866609
 },
 {
-"title": "Domitia cervina",
-"pageid": 52146841
+"title": "The Limited Series (2005 album)",
+"pageid": 3554861
 }
 ]
