+++ tests/cli/expected/random_size_limits.txt
@@ -1,5 +1,5 @@
-Ho Kwan Lung
-Cotys IV
-Lavender and Old Lace (film)
-Monkton Up Wimborne Hundred
-El camino del libertario
+André Bolduc
+2022 Sparta Prague Open
+Chiyorichō-nichōme Station
+Beauchamp Falls
+Begu Khel
