+++ tests/cli/expected/search_de.txt
@@ -1,12 +1,13 @@
-Total hits: 504962
+Total hits: 10
+Suggestion: python unterstütztes mehrerau programmierparadigmen
 
-Berlin
-Berlin, Berlin
-Humboldt-Universität zu Berlin
-Freie Universität Berlin
-Ost-Berlin
-Technische Universität Berlin
-Berliner Mauer
-West-Berlin
-Berlin in Berlin
-Berliner Zeitung
+Python (Programmiersprache)
+R (Programmiersprache)
+JavaScript
+Hybridsprache
+Ruby (Programmiersprache)
+Scratch (Programmiersprache)
+Rust (Programmiersprache)
+Wolfram Language
+Objektorientierte Programmierung
+Tcl
