+++ tests/cli/expected/search_en.txt
@@ -1,12 +1,12 @@
-Total hits: 6043
+Total hits: 6044
 
 Python (programming language)
 History of Python
 Outline of the Python programming language
 Python syntax and semantics
 Mojo (programming language)
-Core Python Programming
-Flask (web framework)
 Zen of Python
-Glob (programming)
 Python
+List of Python software
+Glob (programming)
+Anaconda (Python distribution)
