+++ tests/cli/expected/search_en_json.txt
@@ -1,13 +1,13 @@
 {
-"totalhits": 6043,
+"totalhits": 6044,
 "pages": [
 {
 "title": "Python (programming language)",
 "pageid": 23862,
 "snippet": "<span class=\"searchmatch\">Python</span> is a high-level, general-purpose <span class=\"searchmatch\">programming</span> language. Its design philosophy emphasizes code readability with the use of significant indentation",
-"size": 140415,
+"size": 140424,
 "wordcount": 11448,
-"timestamp": "2026-03-25T15:46:00Z"
+"timestamp": "2026-03-30T16:24:18Z"
 },
 {
 "title": "History of Python",
@@ -42,22 +42,6 @@
 "timestamp": "2026-03-15T08:21:22Z"
 },
 {
-"title": "Core Python Programming",
-"pageid": 25061839,
-"snippet": "Core <span class=\"searchmatch\">Python</span> <span class=\"searchmatch\">Programming</span> is a textbook on the <span class=\"searchmatch\">Python</span> <span class=\"searchmatch\">programming</span> language, written by Wesley J. Chun. The first edition of the book was released on December",
-"size": 3926,
-"wordcount": 316,
-"timestamp": "2026-01-31T19:07:48Z"
-},
-{
-"title": "Flask (web framework)",
-"pageid": 29040606,
-"snippet": "framework and supports <span class=\"searchmatch\">Python</span> 2.7 and 3.5 and later. Jinja, also by Ronacher, is a template engine for the <span class=\"searchmatch\">Python</span> <span class=\"searchmatch\">programming</span> language. Similar to the",
-"size": 10240,
-"wordcount": 831,
-"timestamp": "2026-01-05T17:44:23Z"
-},
-{
 "title": "Zen of Python",
 "pageid": 46448252,
 "snippet": "<span class=\"searchmatch\">Python</span> is a collection of 19 &quot;guiding principles&quot; for writing computer <span class=\"searchmatch\">programs</span> that influence the design of the <span class=\"searchmatch\">Python</span> <span class=\"searchmatch\">programming</span> language. <span class=\"searchmatch\">Python</span> code",
@@ -74,12 +58,28 @@
 "timestamp": "2026-03-10T10:36:25Z"
 },
 {
+"title": "List of Python software",
+"pageid": 3673376,
+"snippet": "The <span class=\"searchmatch\">Python</span> <span class=\"searchmatch\">programming</span> language is actively used by many people, both in industry and academia, for a wide variety of purposes. Atom — an open-source cross-platform",
+"size": 35231,
+"wordcount": 3675,
+"timestamp": "2026-03-27T01:44:00Z"
+},
+{
 "title": "Glob (programming)",
 "pageid": 484117,
 "snippet": "on these same wildcard patterns. Guido van Rossum, author of the <span class=\"searchmatch\">Python</span> <span class=\"searchmatch\">programming</span> language, wrote and contributed a glob routine to BSD Unix in 1986",
 "size": 27549,
-"wordcount": 2505,
+"wordcount": 2499,
 "timestamp": "2026-03-01T01:55:39Z"
+},
+{
+"title": "Anaconda (Python distribution)",
+"pageid": 43003632,
+"snippet": "science and artificial intelligence distribution platform for the <span class=\"searchmatch\">Python</span> <span class=\"searchmatch\">programming</span> language. Developed by Anaconda, Inc., an American company founded",
+"size": 24837,
+"wordcount": 1961,
+"timestamp": "2026-03-30T14:00:32Z"
 }
 ]
 }
