+++ tests/cli/expected/search_sort_timestamp.txt
@@ -1,4 +1,4 @@
-Total hits: 6043
+Total hits: 6044
 
 Programming Without Coding Technology (software)
 AMR (R package)
