+++ tests/cli/expected/search_sort_title.txt
@@ -1,4 +1,4 @@
-Total hits: 6043
+Total hits: 6044
 
 .py
 "Hello, World!" program
