+++ tests/cli/expected/version.txt
@@ -1,1 +1,1 @@
-wikipedia-api, version 0.11.0
+wikipedia-api, version 0.13.0
