+++ tests/cli_test.py
@@ -8,34 +8,34 @@ import pytest
 
 from tests.mock_data import create_mock_wikipedia
 import wikipediaapi
-from wikipediaapi.cli import create_wikipedia_instance
-from wikipediaapi.cli import fetch_page
-from wikipediaapi.cli import format_category_members
-from wikipediaapi.cli import format_coordinates
-from wikipediaapi.cli import format_geosearch
-from wikipediaapi.cli import format_langlinks
-from wikipediaapi.cli import format_page_dict
-from wikipediaapi.cli import format_page_info
-from wikipediaapi.cli import format_random
-from wikipediaapi.cli import format_search
-from wikipediaapi.cli import format_sections
-from wikipediaapi.cli import get_category_members
-from wikipediaapi.cli import get_geosearch_results
-from wikipediaapi.cli import get_langlinks
-from wikipediaapi.cli import get_page_backlinks
-from wikipediaapi.cli import get_page_categories
-from wikipediaapi.cli import get_page_coordinates
-from wikipediaapi.cli import get_page_images
-from wikipediaapi.cli import get_page_info
-from wikipediaapi.cli import get_page_links
-from wikipediaapi.cli import get_page_sections
-from wikipediaapi.cli import get_page_summary
-from wikipediaapi.cli import get_page_text
-from wikipediaapi.cli import get_random_pages
-from wikipediaapi.cli import get_search_results
-from wikipediaapi.cli import get_section_text
-from wikipediaapi.cli import PageNotFoundError
-from wikipediaapi.cli import SectionNotFoundError
+from wikipediaapi.commands.base import create_wikipedia_instance
+from wikipediaapi.commands.base import fetch_page
+from wikipediaapi.commands.base import format_page_dict
+from wikipediaapi.commands.base import format_page_info
+from wikipediaapi.commands.base import format_sections
+from wikipediaapi.commands.base import PageNotFoundError
+from wikipediaapi.commands.base import SectionNotFoundError
+from wikipediaapi.commands.category_commands import format_category_members
+from wikipediaapi.commands.category_commands import get_category_members
+from wikipediaapi.commands.category_commands import get_page_categories
+from wikipediaapi.commands.geo_commands import format_coordinates
+from wikipediaapi.commands.geo_commands import format_geosearch
+from wikipediaapi.commands.geo_commands import get_geosearch_results
+from wikipediaapi.commands.geo_commands import get_page_coordinates
+from wikipediaapi.commands.geo_commands import get_page_images
+from wikipediaapi.commands.link_commands import format_langlinks
+from wikipediaapi.commands.link_commands import get_langlinks
+from wikipediaapi.commands.link_commands import get_page_backlinks
+from wikipediaapi.commands.link_commands import get_page_links
+from wikipediaapi.commands.page_commands import get_page_info
+from wikipediaapi.commands.page_commands import get_page_sections
+from wikipediaapi.commands.page_commands import get_page_summary
+from wikipediaapi.commands.page_commands import get_page_text
+from wikipediaapi.commands.page_commands import get_section_text
+from wikipediaapi.commands.search_commands import format_random
+from wikipediaapi.commands.search_commands import format_search
+from wikipediaapi.commands.search_commands import get_random_pages
+from wikipediaapi.commands.search_commands import get_search_results
 
 
 class TestCLIFactoryFunctions:
