+++ wikipediaapi/commands/__init__.py
@@ -1,0 +1,4 @@
+r"""CLI package for Wikipedia-API.
+
+This package contains the modular command line interface components.
+"""
